requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
insightface==0.7.3
onnxruntime==1.18.1
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
python-multipart==0.0.9
